+++ src/index.tsx
@@ -10,6 +10,21 @@ import Appodeal from './RNAppodeal';
 export { default as AppodealBanner } from './RNAppodealBanner';
 export { default as AppodealMrec } from './RNAppodealMrec';
 export { default as AppodealNative } from './RNAppodealNative';
+export {
+  AppodealNativeAdView,
+  AppodealNativeAsset,
+  AppodealNativeMediaView,
+  AppodealNativeIconView,
+  AppodealNativeTitleView,
+  AppodealNativeDescriptionView,
+  AppodealNativeCallToActionView,
+  AppodealNativeAttributionView,
+} from './RNAppodealNativeAdView';
+export type {
+  AppodealNativeAdViewProps,
+  AppodealNativeAssetProps,
+  AppodealNativeAssetType,
+} from './RNAppodealNativeAdView';
 
 // Export the main Appodeal interface
 export default Appodeal;